literals working in interpreter

+++ src/test/java/ParserUnitTests.java
@@ -79,7 +79,7 @@ public class ParserUnitTests {
         Type type = defOne.getType().getType();
         Position position = defOne.getPosition();
         LiteralInteger integer = (LiteralInteger) defOne.getExpression();
-        int value = integer.getValue();
+        int value = (int) integer.getValue();
         Assert.assertEquals(name, "x");
         Assert.assertEquals(type, Type.INT);
         Assert.assertEquals(position.getLine(), 0);
@@ -90,9 +90,9 @@ public class ParserUnitTests {
         ReturnStatement returnStatement = (ReturnStatement) functionDefinitionList.get("function").getBody().getInstructions().get(1);
         ArthmeticExpression arthmeticExpression = (ArthmeticExpression) returnStatement.getExpression();
         LiteralInteger left = (LiteralInteger) arthmeticExpression.getLeft();
-        int leftValue = left.getValue();
+        int leftValue = (int) left.getValue();
         LiteralInteger right = (LiteralInteger) arthmeticExpression.getRight();
-        int rightValue = right.getValue();
+        int rightValue = (int) right.getValue();
         Assert.assertEquals(leftValue, 2);
         Assert.assertEquals(rightValue, 3);
